+++ Day5/Assignment3/GeneralEmployee.java
@@ -1,7 +1,7 @@
 package Day5.Assignment3;
 
 public class GeneralEmployee extends BaseEmployee {
-    public GeneralEmployee(String empCode, String firstName, String lastName, String empType, double basicSalary) {
+    public GeneralEmployee(String empCode, String firstName, String lastName, String empType, long basicSalary) {
         super(empCode, firstName, lastName, empType, basicSalary);
     }
 }
